+++ examples/plot_power_terms.py
@@ -16,9 +16,9 @@ from thermohl.power import olla
 def plot_joule_heating(dic):
     mdl = [
         dict(label='cigre', model=cigre.JouleHeating(**dic)),
+        dict(label='cner', model=cner.JouleHeating(**dic)),
         dict(label='ieee', model=ieee.JouleHeating(**dic)),
         dict(label='olla', model=olla.JouleHeating(**dic)),
-        dict(label='cner', model=cner.JouleHeating(**dic)),
     ]
     plt.figure()
     t = np.linspace(0., 200., 4001)
@@ -34,11 +34,11 @@ def plot_joule_heating(dic):
 
 
 def plot_solar_heating(dic):
-    # olla not tested here since olla's solar heating is the same as ieee's one
     mdl = [
         dict(label='cigre', cm=cm.winter, model=cigre.SolarHeating(**dic)),
-        dict(label='ieee', cm=cm.autumn, model=ieee.SolarHeating(**dic)),
         dict(label='cner', cm=cm.summer, model=cner.SolarHeating(**dic)),
+        dict(label='ieee', cm=cm.autumn, model=ieee.SolarHeating(**dic)),
+        dict(label='olla', cm=cm.spring, model=olla.SolarHeating(**dic)),
     ]
 
     # examples 1
@@ -97,8 +97,8 @@ def plot_convective_cooling(dic):
     # olla not tested here since olla's convective cooling is the same as ieee's one
     mdl = [
         dict(label='cigre', cm=cm.winter, model=cigre.ConvectiveCooling(**dic)),
-        dict(label='ieee', cm=cm.autumn, model=ieee.ConvectiveCooling(**dic)),
         dict(label='cner', cm=cm.summer, model=cner.ConvectiveCooling(**dic)),
+        dict(label='ieee', cm=cm.autumn, model=ieee.ConvectiveCooling(**dic)),
     ]
 
     fig, ax = plt.subplots(nrows=len(ws), ncols=len(wa))
@@ -169,7 +169,7 @@ if __name__ == '__main__':
 
     dic = solver.default_values()
     dic['lat'] = 45.
-    dic['tb'] = 0.
+    dic['tb'] = 0.33
     dic['al'] = 0.25
 
     plot_joule_heating(dic)
