+++ src/thermohl/power/cner.py
@@ -7,52 +7,126 @@ from typing import Union
 import numpy as np
 
 from thermohl import air
-from thermohl import sun
-from thermohl.power import olla, ieee
+from thermohl.power import ieee
 from thermohl.power.base import PowerTerm
 
 
-class JouleHeating(olla.JouleHeating):
+class JouleHeating(PowerTerm):
     """Joule heating term."""
-    pass
 
+    def __init__(
+            self,
+            I: Union[float, np.ndarray],
+            D: Union[float, np.ndarray],
+            d: Union[float, np.ndarray],
+            A: Union[float, np.ndarray],
+            a: Union[float, np.ndarray],
+            km: Union[float, np.ndarray],
+            ki: Union[float, np.ndarray],
+            kl: Union[float, np.ndarray],
+            kq: Union[float, np.ndarray],
+            RDC20: Union[float, np.ndarray],
+            l: Union[float, np.ndarray],
+            T20: Union[float, np.ndarray] = 20.,
+            f: Union[float, np.ndarray] = 50.,
+            **kwargs
+    ):
+        r"""Init with args.
 
-class SolarHeating(PowerTerm):
-    """Solar heating term.
+        If more than one input are numpy arrays, they should have the same size.
 
-    Very similar to IEEE. Differences explained in methods' comments.
-    """
+        Parameters
+        ----------
+        I : float or np.ndarray
+            Transit intensity.
+        D : float or np.ndarray
+            External diameter.
+        d : float or np.ndarray
+            core diameter.
+        A : float or np.ndarray
+            External (total) section.
+        a : float or np.ndarray
+            core section.
+        km : float or np.ndarray
+            Coefficient for magnetic effects.
+        ki : float or np.ndarray
+            Coefficient for magnetic effects.
+        kl : float or np.ndarray
+            Linear resistance augmentation with temperature.
+        kq : float or np.ndarray
+            Quadratic resistance augmentation with temperature.
+        RDC20 : float or np.ndarray
+            Electric resistance per unit length (DC) at 20°C.
+        l : float or np.ndarray
+            ...
+        T20 : float or np.ndarray, optional
+            Reference temperature. The default is 20.
+        f : float or np.ndarray, optional
+            Current frequency (Hz). The default is 50.
 
-    @staticmethod
-    def _catm(x: Union[float, np.ndarray]) -> Union[float, np.ndarray]:
-        """Compute coefficient for atmosphere turbidity.
-
-        Differences with IEEE version is clean air only and slightly different
-        coefficients.
         """
-        return np.maximum(
-            -42. + 63.8 * x - 1.922 * x**2 + 0.03469 * x**3 - 3.61E-04 * x**4 + 1.943E-06 * x**5 - 4.08E-09 * x**6,
-            0.
-        )
+        self.I = I
+        self.D = D
+        self.d = d
+        self.kem = self._kem(A, a, km, ki)
+        self.kl = kl
+        self.kq = kq
+        self.RDC20 = RDC20
+        self.T20 = T20
+        self.f = f
 
-    @staticmethod
-    def _solar_radiation(
-            lat: Union[float, np.ndarray],
-            azm: Union[float, np.ndarray],
-            month: Union[int, np.ndarray[int]],
-            day: Union[int, np.ndarray[int]],
-            hour: Union[float, np.ndarray]
-    ) -> Union[float, np.ndarray]:
-        """Compute solar radiation.
+    def _rdc(self, T: Union[float, np.ndarray]) -> Union[float, np.ndarray]:
+        """Compute resistance per unit length for direct current."""
+        dt = (T - self.T20)
+        return self.RDC20 * (1. + self.kl * dt + self.kq * dt**2)
 
-        Difference with IEEE version are neither turbidity or altitude influence.
+    def _ks(self, rdc):
+        """Compute skin-effect coefficient."""
+        # Note: approx version as in [NT-RD-CNER-DL-SLA-20-00215]
+        z = 8 * np.pi * self.f * (self.D - self.d)**2 / ((self.D**2 - self.d**2) * 1.0E+07 * rdc)
+        a = 7 * z**2 / (315 + 3 * z**2)
+        b = 56 / (211 + z**2)
+        beta = 1. - self.d / self.D
+        return 1 + a * (1. - 0.5 * beta - b * beta**2)
+
+    def _kem(self, A, a, km, ki) -> Union[float, np.ndarray]:
+        """Compute magnetic coefficient."""
+        s = np.ones_like(self.I) * np.ones_like(A) * np.ones_like(a) * np.ones_like(km) * np.ones_like(ki)
+        z = s.shape == ()
+        if z:
+            s = np.array([1.])
+        I_ = self.I * s
+        a_ = a * s
+        A_ = A * s
+        m = a_ > 0.
+        ki_ = ki * s
+        kem = km * s
+        kem[m] += ki_[m] * I_[m] / ((A_[m] - a_[m]) * 1.0E+06)
+        if z:
+            kem = kem[0]
+        return kem
+
+    def value(self, T: Union[float, np.ndarray]) -> Union[float, np.ndarray]:
+        r"""Compute joule heating.
+
+        Parameters
+        ----------
+        T : float or np.ndarray
+            Conductor temperature.
+
+        Returns
+        -------
+        float or np.ndarray
+            Power term value (W.m\ :sup:`-1`\ ).
+
         """
-        sa = sun.solar_altitude(lat, month, day, hour)
-        sz = sun.solar_azimuth(lat, month, day, hour)
-        th = np.arccos(np.cos(sa) * np.cos(sz - azm))
-        Q = SolarHeating._catm(np.rad2deg(sa))
-        return Q * np.sin(th)
+        rdc = self._rdc(T)
+        ks = self._ks(rdc)
+        rac = self.kem * ks * rdc
+        return rac * self.I**2
 
+
+class SolarHeating(ieee.SolarHeatingBase):
     def __init__(
             self,
             lat: Union[float, np.ndarray],
@@ -95,34 +169,16 @@ class SolarHeating(PowerTerm):
             Power term value (W.m\ :sup:`-1`\ ).
 
         """
-        self.alpha = alpha
-        if srad is None:
-            self.srad = SolarHeating._solar_radiation(np.deg2rad(lat), np.deg2rad(azm), month, day, hour)
-        else:
-            self.srad = srad
-        self.D = D
-
-    def value(self, T: Union[float, np.ndarray]) -> np.ndarray:
-        r"""Compute solar heating.
-
-        If more than one input are numpy arrays, they should have the same size.
-
-        Parameters
-        ----------
-        T : float or np.ndarray
-            Conductor temperature.
-
-        Returns
-        -------
-        float or np.ndarray
-            Power term value (W.m\ :sup:`-1`\ ).
-
-        """
-        return self.alpha * self.srad * self.D * np.ones_like(T)
-
-    def derivative(self, T: Union[float, np.ndarray], **kwargs) -> np.ndarray:
-        """Compute solar heating derivative."""
-        return np.zeros_like(T)
+        est = ieee._SRad(
+            [-42., +63.8, -1.922, 0.03469, -3.61E-04, +1.943E-06, -4.08E-09],
+            [0., 0., 0., 0., 0., 0., 0.]
+        )
+        for k in ['alt', 'tb']:
+            kwargs.pop(k)
+        super().__init__(
+            lat=lat, alt=0., azm=azm, tb=0., month=month, day=day, hour=hour, D=D, alpha=alpha, est=est, srad=srad,
+            **kwargs
+        )
 
 
 class ConvectiveCooling(PowerTerm):
